backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
nltk==3.6.5
numpy==1.21.4
gunicorn==20.1.0
python-dotenv==0.19.2
requests==2.28.1
